+++ R/col_vals_decreasing.R
@@ -421,6 +421,7 @@ col_vals_decreasing <- function(
   
   # Add one or more validation steps based on the
   # length of the `columns` variable
+  label <- resolve_label(label, columns, segments_list)
   for (i in seq_along(columns)) {
     for (j in seq_along(segments_list)) {
       
@@ -442,7 +443,7 @@ col_vals_decreasing <- function(
           seg_val = seg_val,
           actions = covert_actions(actions, agent),
           step_id = step_id[i],
-          label = label,
+          label = label[[i, j]],
           brief = brief[i],
           active = active
         )