+++ R/rows_complete.R
@@ -352,6 +352,7 @@ rows_complete <- function(
   
   # Add one or more validation steps based on the
   # length of `segments`
+  label <- resolve_label(label, segments = segments_list)
   for (i in seq_along(segments_list)) {
     
     seg_col <- names(segments_list[i])
@@ -371,7 +372,7 @@ rows_complete <- function(
         seg_val = seg_val,
         actions = covert_actions(actions, agent),
         step_id = step_id,
-        label = label,
+        label = label[[i]],
         brief = brief,
         active = active
       )