+++ R/rows_distinct.R
@@ -353,6 +353,7 @@ rows_distinct <- function(
   
   # Add one or more validation steps based on the
   # length of `segments`
+  label <- resolve_label(label, segments = segments_list)
   for (i in seq_along(segments_list)) {
     
     seg_col <- names(segments_list[i])
@@ -372,7 +373,7 @@ rows_distinct <- function(
         seg_val = seg_val,
         actions = covert_actions(actions, agent),
         step_id = step_id,
-        label = label,
+        label = label[[i]],
         brief = brief,
         active = active
       )