+++ R/steps_and_briefs.R
@@ -35,7 +35,8 @@ create_validation_step <- function(
     step_id = NULL,
     label = NULL,
     brief = NULL,
-    active = NULL
+    active = NULL,
+    .call = rlang::caller_env(n = 2L)
 ) {
   
   # Get the next step number (i)
@@ -105,6 +106,21 @@ create_validation_step <- function(
       f_passed = NA_real_,
       f_failed = NA_real_
     )
+  
+  # glue-powered labels
+  if (!is.na(validation_step_df$label)) {
+    glue_mask <- rlang::new_environment(
+      data = list(
+        .step = assertion_type,
+        .col = column,
+        .seg_col = seg_col,
+        .seg_val = seg_val
+      ),
+      parent = .call
+    )
+    label <- glue::glue_data(glue_mask, validation_step_df$label, .envir = NULL)
+    validation_step_df$label <- as.character(label)
+  }
   
   # Append `validation_step` to `validation_set`
   agent$validation_set <-